adding error messaging for unsupported tensor shapes
