+++ monai/csrc/filtering/bilateral/bilateralfilter_cuda_phl.cu
@@ -15,6 +15,7 @@ limitations under the License.
 #include <cuda_runtime.h>
 #include <torch/extension.h>
 
+#include "bilateral.h"
 #include "filtering/permutohedral/permutohedral.h"
 #include "utils/meta_macros.h"
 #include "utils/tensor_description.h"
@@ -135,7 +136,7 @@ torch::Tensor BilateralFilterPHLCuda(torch::Tensor inputTensor, float spatialSig
                                    inputTensor, outputTensor, spatialSigma, colorSigma); \
                              }));
 
-  SWITCH_AB(CASE, 16, 3, inputTensor.size(1), inputTensor.dim() - 2);
+  SWITCH_AB(CASE, BF_CUDA_MAX_CHANNELS, BF_CUDA_MAX_SPATIAL_DIMENSION, inputTensor.size(1), inputTensor.dim() - 2);
 
   return outputTensor;
 }
