+++ monai/csrc/filtering/permutohedral/permutohedral_cuda.cu
@@ -38,7 +38,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#define BLOCK_SIZE 64
+#define BLOCK_SIZE 32
 
 #include <cuda.h>
 #include <cuda_runtime.h>
@@ -46,6 +46,7 @@ SOFTWARE.
 #include <torch/extension.h>
 #include <THC/THCAtomics.cuh>
 
+#include "permutohedral.h"
 #include "hash_table.cuh"
 #include "utils/meta_macros.h"
